Modified application
Proposed change of the top bar where now it goes all the way across, still has the hamburger icon, the bar is fixed at the top, when pressed it doesn't force you to to the top, shows the offcanvas menu where you're located so you can go back to your original location, and displays GIFWAR on the right side of the bar.

A buffer still needs to be placed so that the fixed bar doesn't overlap so if need be we can go back to just the icon.

Modified the top bar to be sticky

Removed trailing whtespace

Modified application

Proposed change of the top bar where now it goes all the way across, still has the hamburger icon, the bar is fixed at the top, when pressed it doesn't force you to to the top, shows the offcanvas menu where you're located so you can go back to your original location, and displays GIFWAR on the right side of the bar.

A buffer still needs to be placed so that the fixed bar doesn't overlap so if need be we can go back to just the icon.